[mkcal] Don't insert the organizer as an attendee.
This modification creates problems when importing and exporting
incidences, as e.g. the list of attendees in the actual inserted event
will be different from the one passed to insertEvent().

Also, if there were previously no attendees for an event, then adding
the organizer as an attendee turns the event into a calendar
invitation with RSVP details, even if it is not intended to be one.